feat(posh): add static data
This commit adds static data to the Posh target. It includes a new class, `PoshStatic`, which implements the `StaticData` interface, as well as changes to the `Posh` class to remove unused imports and test methods. The `PoshStatic` class provides lists of JARM fingerprints, certificates, server ports, and HTTP headers, and SHA-256 hashes of response bodies.